apps/chatui/pyproject.toml
[project]
name = "chatui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp-sse-client>=0.2.1",
    "chainlit>=2.5.5",
    "google-adk>=1.0.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "sseclient-py>=1.8.0",
]
